requirements.txt
 azure-cognitiveservices-vision-face
 azure-identity